core toolchain packages should also get scheduled. To properly separate out scheduler and builder code so that our builder component be it thread or container/pod should just ask for what package they need to build and build it.
Currently coretool chain packages are not scheduled.
They are build sequentially as defined in constants.py

To make sure build happens in the same sequence manner as currently, scheduler should
schedule them in the same way.
For this to happend we make sure packages which are build first are
child node of the rest of the remaining core toolchain packages.
Hence if they are not build the rest of core toolchain build cannot happen.

Eg: filesystem is child node of each of the packages. So unless it is build
no other package is scheduled.

linux-api-header is the child node of rest of packages to right of the
constants.listCoreToolChainPackages list. That means unless linux-api-header is build
none of the other remaining packages are scheduled to build

Change-Id: Ie38bd15b8ef29379b07f1d7642e1a6bc9d7ab24b